+++ src/sys/unix/waker.rs
@@ -118,36 +118,7 @@ pub(crate) use self::eventfd::WakerInternal;
         target_os = "watchos",
     )
 ))]
-mod kqueue {
-    use crate::sys::Selector;
-    use crate::Token;
-
-    use std::io;
-
-    /// Waker backed by kqueue user space notifications (`EVFILT_USER`).
-    ///
-    /// The implementation is fairly simple, first the kqueue must be setup to
-    /// receive waker events this done by calling `Selector.setup_waker`. Next
-    /// we need access to kqueue, thus we need to duplicate the file descriptor.
-    /// Now waking is as simple as adding an event to the kqueue.
-    #[derive(Debug)]
-    pub struct Waker {
-        selector: Selector,
-        token: Token,
-    }
-
-    impl Waker {
-        pub fn new(selector: &Selector, token: Token) -> io::Result<Waker> {
-            let selector = selector.try_clone()?;
-            selector.setup_waker(token)?;
-            Ok(Waker { selector, token })
-        }
-
-        pub fn wake(&self) -> io::Result<()> {
-            self.selector.wake(self.token)
-        }
-    }
-}
+mod kqueue;
 
 #[cfg(all(
     not(mio_unsupported_force_waker_pipe),
